requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
SQLAlchemy==2.0.30
Werkzeug==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Pillow==10.3.0
